add testutils.CaptureOutput() - from stormkit.captureOutput()
