+++ stormkit/main_test.go
@@ -1,16 +1,13 @@
 package stormkit
 
 import (
-	"bytes"
 	"errors"
-	"io"
-	"log"
 	"os"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/spf13/viper"
+	"github.com/stormkit-io/stormkit-cli/testutils"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -76,7 +73,7 @@ app:
   - id: ` + expectedAppID), nil
 	}
 
-	re := captureOutput(func() {
+	re := testutils.CaptureOutput(func() {
 		ConfigWithPath(".")
 	})
 
@@ -99,42 +96,12 @@ app:
   - id: "ccccc"`), nil
 	}
 
-	re := captureOutput(func() {
+	re := testutils.CaptureOutput(func() {
 		Config()
 		ConfigWithPath("./test")
 	})
 
 	assert.Equal(t, "there are many apps in the config file (using the first)\n", re)
-}
-
-func captureOutput(f func()) string {
-	reader, writer, err := os.Pipe()
-	if err != nil {
-		panic(err)
-	}
-	stdout := os.Stdout
-	stderr := os.Stderr
-	defer func() {
-		os.Stdout = stdout
-		os.Stderr = stderr
-		log.SetOutput(os.Stderr)
-	}()
-	os.Stdout = writer
-	os.Stderr = writer
-	log.SetOutput(writer)
-	out := make(chan string)
-	wg := new(sync.WaitGroup)
-	wg.Add(1)
-	go func() {
-		var buf bytes.Buffer
-		wg.Done()
-		io.Copy(&buf, reader)
-		out <- buf.String()
-	}()
-	wg.Wait()
-	f()
-	writer.Close()
-	return <-out
 }
 
 func TestGetStormkitConfigFilePathNotExits(t *testing.T) {
